+++ server/lib/src/config/config.dart
@@ -15,4 +15,12 @@ abstract class Config {
   /// In how many seconds white has to perform a move before the game is
   /// disbanded.
   static const int secondsWhiteForFirstMove = 30;
+
+  /// The duration in which a private challenge needs to be accepted before it
+  /// is cancelled by the server.
+  static const Duration expirePrivateChallengeAfter = Duration(minutes: 2);
+
+  /// The duration in which a public challenge needs to be matchmade before it
+  /// is cancelled by the server.
+  static const Duration expirePublicChallengeAfter = Duration(hours: 1);
 }